+++ database/migrations/2023_12_10_090138_create_raw_materials_groups_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->bigInteger('raw_materials_id')->nullable();
             $table->string('group_name')->nullable();
-            $table->bigInteger('grams')->nullable();
+            $table->float('grams')->nullable();
             $table->bigInteger('token')->nullable();
             $table->timestamps();
         });
